Extract Horde Settings
